backend/requirements.txt
numpy==1.24.3
pandas==2.0.3
pandas-ta==0.3.14b0
python-binance==1.0.19
flask==3.0.2
flask-cors==4.0.0